数据字典，职位，系统配置，部门

+++ mydemo/src/main/java/com/wy/mydemo/mapper/SysUserPostMapper.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 @Mapper
 @Repository
 public interface SysUserPostMapper {
-    long countByExample(SysUserPostExample example);
+    int countByExample(SysUserPostExample example);
 
     int deleteByExample(SysUserPostExample example);
